Change AppInsights `type` in the manifest
Before Preview 1 of Aspire and `azd` we made a change such that all
the resource types were `.v0` instead of `.v1`.

That change didn't happen in the manifest writer for the App Insights
resource because we only did passes over the Aspire and `azd`
repositories.

This fixes the manifest writer to write the type that `azd` expects.
With this change you can now `azd init` and `azd provision` this
sample.